Rename subscribeable -> subscribable
